Move regression test to correct location per reviewer feedback
dmiltr3 requested the test be placed in:
  tensorflow/python/kernel_tests/io_ops/parsing_ops_test.py
instead of the incorrectly created:
  tensorflow/python/ops/parsing_ops_test.py

Updated test to use project-standard imports (parsing_ops, dtypes)
instead of tf.io and tf.variant. Class appended before if __name__
block following existing file conventions.
